+++ lib/messages.ex
@@ -53,6 +53,8 @@ defmodule FlowAssertions.Messages do
   def wrong_field_value(key), do: "Field `#{inspect key}` has the wrong value"
   def field_missing(field), do: "Field `#{inspect field}` is missing"
 
+  def field_wrongly_present(field), do: "Field `#{inspect field}` should not be present."
+
   def not_enumerable, do: "Expected an `Enumerable`"
 
   def different_length_collections, do: "The two collections have different lengths"